+++ history.md
@@ -1,6 +1,7 @@
 # 版本信息
 ## 2.16.0 (2018-01-03)
 * [ADD] 新增 `yyl server rebuild <projectName>`
+* [TODO] 执行 watch 后 修改 yy.com 下的 w-thirdpartyLogin.js 不会触发更新 js 操作
 
 ## 2.15.28 (2017-12-28)
 * [ADD] `yyl init` 新增 `--name`, `--platform`, `--workflow`, `--init`, `--doc` 参数